Ajustando os testes de vez
