+++ frontend/src/entities/__tests__/Evolucao.spec.js
@@ -2,9 +2,7 @@ import Evolucao from '@/entities/Evolucao';
 
 jest.mock('moment', () => {
   return () => ({
-    format: () => {
-      return '18/04/2020 12:53';
-    },
+    format: () => '18/04/2020 12:53',
   });
 });
