+++ frontend/src/entities/__tests__/NotificacaoEvolucao.spec.js
@@ -34,14 +34,6 @@ const notifEvolucaoRequestValida = {
   tpEvolucao: 'SUSPEITO',
 };
 
-const notifEvolucaoRequestBodyValida = {
-  id: 'd67c4ef1-eed6-4c98-b53d-f050a76dfb3a',
-  notificacaoId: 'cd6286f5-dc11-4932-98e3-5ff1893885ab',
-  dataHoraAtualizacao: '18/04/2020 15:53',
-  local: 'Isolamento domiciliar',
-  situacao: 'Suspeito',
-};
-
 describe('Testes para entidade NotificacaoEvolucao', () => {
   test('Criando nova Notificação Evolução', () => {
     const result = new NotificacaoEvolucao();
@@ -58,15 +50,5 @@ describe('Testes para entidade NotificacaoEvolucao', () => {
     dataRequest.dtEvolucao = '18/04/2020 18:53';
     const result = new NotificacaoEvolucao(dataRequest).toRequest();
     expect(result).toEqual(notifEvolucaoRequestValida);
-  });
-
-  test('Criando nova Notificação Evolução com request body', () => {
-    jest.mock('moment', () => {
-      return () => ({
-        format: () => '18/04/2020 15:53',
-      });
-    });
-    const result = new NotificacaoEvolucao(dataNotifEvolucao).toRequestBody();
-    expect(result).toEqual(notifEvolucaoRequestBodyValida);
   });
 });
